13th commit:(Search & Finish) Reports Page

+++ ShahidSoltaniLibrary.App/ReportsForm.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement;
 
 namespace ShahidSoltaniLibrary.App
 {
@@ -21,9 +22,9 @@ namespace ShahidSoltaniLibrary.App
             UpdateData();
         }
 
-        private void UpdateData()
+        private void UpdateData(string query = "", string status = "")
         {
-            grd.DataSource = _uow.LoanService.GetAllLoan(txtquery.Text, cmd.SelectedItem?.ToString());
+            grd.DataSource = _uow.LoanService.GetAllLoan(query, status);
         }
 
         private void btnClose_Click(object sender, EventArgs e)
@@ -46,23 +47,26 @@ namespace ShahidSoltaniLibrary.App
 
         private void grd_CellClick(object sender, DataGridViewCellEventArgs e)
         {
-            btnEdit.Enabled = true;
             btnEnd.Enabled = true;
             btnDelete.Enabled = true;
         }
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            if (grd.SelectedCells != null)
+            var status = MessageBox.Show("آیا مطمئنی؟", "d", MessageBoxButtons.OKCancel);
+            if (grd.SelectedCells != null && status == DialogResult.OK)
             {
                 int loanId = Convert.ToInt32(grd.SelectedCells[0].Value);
                 var loan = _uow.LoanService.GetLoanById(loanId);
 
-                foreach (var ubook in loan.UserBooks)
+                if (!loan.Finish)
                 {
-                    var book = _uow.BookService.GetBookById(ubook.BookId);
-                    book.RemainNumber++;
-                    _uow.BookService.Update(book);
+                    foreach (var ubook in loan.UserBooks)
+                    {
+                        var book = _uow.BookService.GetBookById(ubook.BookId);
+                        book.RemainNumber++;
+                        _uow.BookService.Update(book);
+                    }
                 }
                 _uow.LoanService.Delete(loan);
                 bool res = _uow.Save();
@@ -74,6 +78,43 @@ namespace ShahidSoltaniLibrary.App
 
                 UpdateData();
             }
+        }
+
+        private void btnEnd_Click(object sender, EventArgs e)
+        {
+            if (grd.SelectedCells != null)
+            {
+                var status = MessageBox.Show("آیا مطمئنی؟", "d", MessageBoxButtons.OKCancel);
+                int loanId = Convert.ToInt32(grd.SelectedCells[0].Value);
+                var loan = _uow.LoanService.GetLoanById(loanId);
+                if (status == DialogResult.OK && loan.Finish)
+                {
+                    MessageBox.Show("عملیات از قبل انجام شده");
+                    return;
+                }
+                loan.Finish = true;
+                loan.EndDate = DateTime.Now;
+                _uow.LoanService.Update(loan);
+                foreach (var ubook in loan.UserBooks)
+                {
+                    var book = _uow.BookService.GetBookById(ubook.BookId);
+                    book.RemainNumber++;
+                    _uow.BookService.Update(book);
+                }
+                bool res = _uow.Save();
+
+                if (res)
+                    MessageBox.Show("موفقیت");
+                else
+                    MessageBox.Show("شکست");
+
+                UpdateData();
+            }
+        }
+
+        private void bunifuButton21_Click(object sender, EventArgs e)
+        {
+            UpdateData(txtquery.Text, cmd.SelectedItem?.ToString());
         }
     }
 }
